+++ diesel_codegen_syntex/src/identifiable.rs
@@ -4,6 +4,7 @@ use syntax::ext::base::{Annotatable, ExtCtxt};
 use syntax::parse::token::str_to_ident;
 
 use model::Model;
+use util::lifetime_list_tokens;
 
 pub fn expand_derive_identifiable(
     cx: &mut ExtCtxt,
@@ -15,12 +16,14 @@ pub fn expand_derive_identifiable(
     if let Some(model) = Model::from_annotable(cx, span, annotatable) {
         let table_name = model.table_name();
         let struct_ty = &model.ty;
+        let lifetimes = lifetime_list_tokens(&model.generics.lifetimes, span);
         let fields = model.field_tokens_for_stable_macro(cx);
         if model.attr_named(str_to_ident("id")).is_some() {
             push(Annotatable::Item(quote_item!(cx, Identifiable! {
                 (
                     table_name = $table_name,
                     struct_ty = $struct_ty,
+                    lifetimes = ($lifetimes),
                 ),
                 fields = [$fields],
             }).unwrap()));